pyproject.toml
[tool.poetry]
name = "codetri-webhook"
version = "0.1.0"
description = ""
authors = ["Caleb Ely <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.0"
watchdog = "^0.9.0"
python-dotenv = "^0.10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
